[refactor] jenkins_job -> jenkins_notification으로 프로젝트 구조 변경
