+++ backend/src/main/java/com/example/backend/jenkins/notification/model/JobNotificationId.java
@@ -1,4 +1,4 @@
-package com.example.backend.jenkins.job.model;
+package com.example.backend.jenkins.notification.model;
 
 import java.io.Serializable;
 import java.time.LocalDateTime;
